chore: rename pool and move format functions to util
