+++ utils/helpers.ts
@@ -1,3 +1,4 @@
+import { FormattedPool, FormattedPosition, RawPool, RawPosition } from '@/hooks/types';
 import { formatUnits } from 'viem';
 const formatNumberWithSuffix = (num: number): string => {
   const formatWithPrecision = (value: number) => {
@@ -56,6 +57,48 @@ function formatAddress(address: string): string {
   return `${address.slice(0, 6)}...${address.slice(-4)}`;
 }
 
+function formatPool(pool: RawPool, isStable: boolean): FormattedPool {
+  return {
+    ...pool,
+    lp: formatAddress(pool.lp),
+    liquidity: formatBigInt(pool.liquidity, pool.decimals),
+    type: pool.type === -1 ? 'CL' : 'V2',
+    stable: isStable,
+    token0: formatAddress(pool.token0),
+    reserve0: formatBigInt(pool.reserve0, pool.decimals),
+    staked0: formatBigInt(pool.staked0, pool.decimals),
+    token1: formatAddress(pool.token1),
+    reserve1: formatBigInt(pool.reserve1, pool.decimals),
+    staked1: formatBigInt(pool.staked1, pool.decimals),
+    gauge: formatAddress(pool.gauge),
+    gauge_liquidity: formatBigInt(pool.gauge_liquidity, pool.decimals),
+    fee: formatAddress(pool.fee),
+    bribe: formatAddress(pool.bribe),
+    factory: formatAddress(pool.factory),
+    emissions: formatBigInt(pool.emissions, 18), // Assuming emissions are in 18 decimals
+    emissions_token: formatAddress(pool.emissions_token),
+    pool_fee: formatPercentage(pool.pool_fee),
+    token0_fees: formatBigInt(pool.token0_fees, pool.decimals),
+    token1_fees: formatBigInt(pool.token1_fees, pool.decimals),
+  };
+}
+
+function formatPosition(position: RawPosition, decimals: number): FormattedPosition {
+  return {
+    id: position.id.toString(),
+    lp: formatAddress(position.lp),
+    liquidity: formatBigInt(position.liquidity, decimals),
+    staked: formatBigInt(position.staked, decimals),
+    amount0: formatBigInt(position.amount0, decimals),
+    amount1: formatBigInt(position.amount1, decimals),
+    staked0: formatBigInt(position.staked0, decimals),
+    staked1: formatBigInt(position.staked1, decimals),
+    unstaked_earned0: formatBigInt(position.unstaked_earned0, decimals),
+    unstaked_earned1: formatBigInt(position.unstaked_earned1, decimals),
+    emissions_earned: formatBigInt(position.emissions_earned, 18), // Assuming emissions are in 18 decimals
+  };
+}
+
 export {
   formatNumberWithSuffix,
   truncateDecimal,
@@ -65,4 +108,6 @@ export {
   formatBigInt,
   formatPercentage,
   formatAddress,
+  formatPool,
+  formatPosition,
 };
